Move from using Double to Float.
This change makes it easier to migrate from Klutter to Crane
as we are going to use Float values rather than Doubles.

Test: ran all existing tests and UI demos

+++ ui/port/src/main/java/androidx/ui/Hooks.kt
@@ -37,17 +37,17 @@ import androidx.ui.engine.window.Window
 import androidx.ui.engine.window.WindowPadding
 
 fun Window.updateWindowMetrics(
-    devicePixelRatio: Double,
-    width: Double,
-    height: Double,
-    paddingTop: Double,
-    paddingRight: Double,
-    paddingBottom: Double,
-    paddingLeft: Double,
-    viewInsetTop: Double,
-    viewInsetRight: Double,
-    viewInsetBottom: Double,
-    viewInsetLeft: Double
+    devicePixelRatio: Float,
+    width: Float,
+    height: Float,
+    paddingTop: Float,
+    paddingRight: Float,
+    paddingBottom: Float,
+    paddingLeft: Float,
+    viewInsetTop: Float,
+    viewInsetRight: Float,
+    viewInsetBottom: Float,
+    viewInsetLeft: Float
 ) {
     this.devicePixelRatio = devicePixelRatio
     this.physicalSize = Size(width, height)
